backend/requirements.txt
Django
psycopg2
sendgrid>=6.9
djangorestframework>=3.12
django-cors-headers>=3.13
redis
celery